move delegation tabbed component into correct folder
