+++ src/broadcast/page.rs
@@ -2,6 +2,8 @@ use super::components::*;
 use crate::common::components::*;
 use leptos::*;
 
+use crate::common::models::*;
+
 #[component]
 pub fn BroadcastTransactionPage() -> impl IntoView {
     view! {
@@ -112,4 +114,30 @@ pub fn BroadcastFromLedgerPage() -> impl IntoView {
             </AppSection>
         </PageContainer>
     }
+}
+
+
+#[component]
+pub fn DelegationTabbedPage() -> impl IntoView {
+    let tabs = vec![
+        NavEntry {
+            href: "/broadcast/transaction".to_string(),
+            text: "Transaction".to_string(),
+            icon: NavIcon::Transactions,
+            sub_entries: None,
+        },
+        NavEntry {
+            href: "/broadcast/delegation".to_string(),
+            text: "Delegation".to_string(),
+            icon: NavIcon::Transactions,
+            sub_entries: None,
+        },
+        NavEntry {
+            href: "/broadcast/ledger".to_string(),
+            text: "Ledger".to_string(),
+            icon: NavIcon::Transactions,
+            sub_entries: None,
+        },
+    ];
+    view! { <TabbedPage tabs/> }
 }
