+++ src/root.rs
@@ -5,9 +5,8 @@ use crate::account_dialog::dialog::AccountDialogView;
 use crate::accounts::page::{AccountSpotlightPage, AccountsPage};
 use crate::blocks::page::{BlockSpotlight, LatestBlocksPage};
 use crate::broadcast::page::{
-    BroadcastDelegationPage, BroadcastFromLedgerPage, BroadcastTransactionPage,
+    BroadcastDelegationPage, BroadcastFromLedgerPage, BroadcastTransactionPage, DelegationTabbedPage
 };
-use crate::common::components::DelegationTabbedPage;
 use crate::footer::Footer;
 use crate::header::navigation::Header;
 use crate::next_stakes::page::NextStakesPage;
